Move creating express instance to a service
